WIP - Another big restructuring. I need to get the structure more easy. Now I removed the api projects and moved them into the client. So I got rid of a complete level of projects.
Additionally moved the contacts and files to the modules. And the
license to base. So the technical modules needed for everything are on
base and everything else is a module. But modules may depend on each
other so contacts (including gis) and files will be default modules for
all others.